Move task related classes into the tasks package
