+++ src/main/java/alvis/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Alvis;
+package alvis.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
