+++ src/main/java/alvis/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package Alvis;
+package alvis.tasks;
 
 public class ToDo extends Task {
     public ToDo(String desc) {
